Starship-bot/requirements.txt
discord.py==2.0.1
Flask==2.2.2
